gui/guiv2/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plasmidvcs"
version = "0.1.0"
description = "Version control for genetic constructs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Igor Sinelnikov", email = "i.sinelnikov@fbras.ru" },
]
keywords = ["bioinformatics", "plasmid", "version-control", "genetic-engineering", "synthetic-biology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "biopython>=1.83",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
fast = [
    "parasail>=1.3",  # faster sequence alignment
]

[project.scripts]
pvcs = "pvcs.cli:main"

[project.urls]
Homepage = "https://github.com/isinelnikov/plasmidvcs"
Issues = "https://github.com/isinelnikov/plasmidvcs/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pvcs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
